feat: set defaults for impossible ForestCentre XML values
Municipality id and forest centre id are plain not available in source data.
Forest management category is now 1 when missing.

Out of two solutions for this, this is the less wrong one. Problem pertains
only to the ForestCentre data source. Problem manifests in RSD format
attempting to encode unknown into a number value space. This could have been
implemented in the RSD data row preparation for ForestStand, but that would
then become a risk source for hiding broken data unwittingly.