chore: static analysis config changes and fixes (#98)
## Updated pylint and mypy configuration and fixed several issues

### mypy
- Added ignore_missing_imports for pymotti
- Added ignore_errors for lukefi.metsi.forestry,
lukefi.metsi.domain.data_collection,
lukefi.metsi.domain.forestry_operations and
lukefi.metsi.domain.natural_processes, essentially disabling their
checks. The detected issues are unlikely to be addressed and at least
large parts of the packages are expected to be rewritten in the near
future.

### pylint
- Set fail-under to 0, making it so the process always returns
succesfully. This is to prevent an unnecessary annotation from the task
failing, when all the specific issues should already generate individual
annotations.
- Set jobs to 0, so that multiple processes are used if cores are
available.
- Un-disabled many checks that did not need to be disabled
- Disabled several checks that generated lots of notice-level issues
that were unlikely to ever be fixed and were of dubious helpfulness to
begin with:
    - duplicate-code
    - too-few-public-methods
    - too-many-arguments
    - too-many-branches
    - too-many-instance-attributes
    - too-many-locals
    - too-many-public-methods
    - too-many-positional-arguments
    - too-many-return-statements
    - too-many-statements
- Explicitly enabled the useless-suppression check to potentially catch
unnecessarily suppressed checks in the future.
- Set ignore-imports to 'yes' for detecting similarities. This has no
effect currently, since the whole check is disabled, but should it be
re-enabled at some point, this would be good to have. Similar looking
lists of imports in different files is not something we should be
worrying about.

### Fixes
- Most error and warning level issues are now fixed
- All notice level issues are either fixed or skipped
- This should allow for potential new issues to be more easily detected,
since they won't be as likely to be left out of the 10 annotation quota.