test, fix: add more robot tests, fix bugs and improve calculation accuracy (#100)
- DryRun for a bare minimum smoketest. Runs "python -m
lukefi.metsi.app.lukefi -h" and checks that the process exits
successfully.
- VMI13_mini for running the default control.py with the
VMI13_source_mini dataset and comparing output.

- Branched earlier versions (1.2.1, 2.0.2, 2.2.0) and reverted the Njit
performance changes while keeping the y1 -> y1t, etc. fix in
'_taper_curve_correction'.
- Ran the tests and saved output to use as reference

$$V =
\frac{\pi}{4}\int_{h_1}^{h_2}\left[\frac{1}{100}\left(ax+bx^2+cx^3+dx^5+ex^8+fx^{13}+gx^{21}+hx^{34}\right)\right]^2
\text{d}\bar{h}$$
$$x = \frac{H-\bar{h}}{H} \Rightarrow \text{d}\bar{h} = -H\text{d}x$$
$$\Rightarrow V = -H
\frac{\pi}{4}\int_{\bar{h}=h_1}^{\bar{h}=h_2}\left[\frac{1}{100}\left(ax+bx^2+cx^3+dx^5+ex^8+fx^{13}+gx^{21}+hx^{34}\right)\right]^2
\text{d}x$$
$$=-H \left.\frac{\pi}{40000}
\left[
\frac{a^2}{3} x^3 +
\frac{ab}{2} x^4 +
\frac{2ac + b^2}{5} x^5 +
\frac{bc}{3} x^6 +
\frac{2ad + c^2}{7} x^7 +
\frac{bd}{4} x^8 +
\frac{2cd}{9} x^9 +
\frac{ae}{5} x^{10} + \right. \right. $$
$$\left. \left.
\frac{2be + d^2}{11} x^{11} +
\frac{ce}{6} x^{12} +
\frac{de}{7} x^{14} +
\frac{2af}{15} x^{15} +
\frac{bf}{8} x^{16} +
\frac{2cf + e^2}{17} x^{17} +
\frac{2df}{19} x^{19} +
\frac{ef}{11} x^{22} +
\frac{2ag}{23} x^{23} + \right. \right. $$
$$\left. \left.
\frac{bg}{12} x^{24} +
\frac{2cg}{25} x^{25} +
\frac{2dg + f^2}{27} x^{27} +
\frac{eg}{15} x^{30} +
\frac{2fg}{35} x^{35} +
\frac{ah}{18} x^{36} +
\frac{2bh}{37} x^{37} +
\frac{ch}{19} x^{38} + \right. \right. $$
$$\left. \left.
\frac{dh}{20} x^{40} +
\frac{2eh + g^2}{43} x^{43} +
\frac{fh}{24} x^{48} +
\frac{gh}{28} x^{56} +
\frac{h^2}{69} x^{69}
\right]
\right|_{\bar{h}=h_1}^{\bar{h}=h_2}$$
- '_volume' now uses the analytic solution for integral when calculating
stem profiles. The original scipy.integrate.quad method was remarkably
accurate, but too slow, while the linear approximation was fast but, as
the addition of more test cases revealed, seemingly not accurate enough.
The analytic solution is as accurate as it gets and also does not seem
to be that much slower than the approximation.
- As a result, it was possible to lower the tolerance of the numeric
comparisons in the test cases all the way down to 0.0000001.

- Fixed the function scan_operation_type_for_event. Bug detected by the
VMI13_mini test. Originally the function had a try/catch block that hid
the bug. The try/catch was removed in a recent cleanup (#95) and the
function was unfortunately not covered by any of the test cases, but
would lead to a crash if used. This should both fix the crash and
actually implement the intended behavior and not just except out with
"unknown_operation".