feature: events and treatments, #MELA2-65 (#114)
## Simulation events are now declared as Event objects in the control
file

Explicit time points are, for the time being, still used for building
the event tree. This is planned to be replaced in the near future by
event and treatment conditions and the transition mechanism.

It should be possible to port all previous control declarations to the
new format. There are however a few differences. Most notably implicit
alternatives for multiple parameter sets for the same operation are no
longer supported, while it is instead possible to declare separate
Treatments with same function but different parameters. Now the same
functionality can be achieved by an explicit Alternatives generator with
the Treatments with different parameters.

Some time should be taken at some point for packaging the template
Treatments in a more friendly way, so that as little boilerplate as
possible is needed in the control files.

Several classes have also been moved from core_types to more bespoke
modules to prevent issues with circular imports.